utilize manual list of httpreleasefeeds, etc
Ordering was screwy - we should be checking for a SUFeedURL(, and may
as well get bundle ID while we're there,) before we worry about the
location of the framework, since all sparkle apps (like devmate ones, I
think) don't put the framework at that path we were expecting. This
errs on the side of 'if I can't find the sparkle path to validate the
version, just lump it in'. Also checking the ~100 apps that use https
feeds but http RELEASE NOTES feeds.